100-atoi.c

+++ pointers_arrays_strings/100-atoi.c
@@ -10,7 +10,7 @@
 int _atoi(char *s)
 {
 int sign = 1;
-unsigned int result = 0;
+long int result = 0;
 
 while (*s)
 {
